requirements.txt
-i https://pypi.org/simple
aniso8601==9.0.1
blinker==1.6.2 ; python_version >= '3.7'
click==8.1.3 ; python_version >= '3.7'
colorama==0.4.6 ; platform_system == 'Windows'
flask==2.3.2
flask-restful==0.3.10
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
markupsafe==2.1.3 ; python_version >= '3.7'
numpy==1.24.3 ; python_version >= '3.10'
pandas==2.0.2
psycopg2-binary==2.9.6
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2023.3
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tzdata==2023.3 ; python_version >= '2'
werkzeug==2.3.4 ; python_version >= '3.8'
